HT_23/sears-celery/requirements/base.txt
celery[redis]==5.3.6
django==5.0
django-extensions==3.2.3
djangorestframework==3.14.0
django-environ==0.11.2
drf-spectacular==0.27.1
loguru==0.7.2
requests==2.31.0